fix: remove unused code and simplify the oob implementation and include attachments
