+++ src/edge-agent/protocols/invitation/v2/OutOfBandInvitation.ts
@@ -1,5 +1,5 @@
 import { uuid } from "@stablelib/uuid";
-import { JsonString } from "../../../../domain";
+import { AttachmentDescriptor, JsonString } from "../../../../domain";
 import { AgentError } from "../../../../domain/models/Errors";
 import { PrismOnboardingInvitation } from "../../../types";
 
@@ -12,8 +12,9 @@ export class OutOfBandInvitation {
   constructor(
     public body: OutOfBandInvitationBody,
     public from: string,
-    public id: string = uuid()
-  ) {}
+    public id: string = uuid(),
+    public attachments: AttachmentDescriptor[] = []
+  ) { }
 
   static parsePrismOnboardingInvitationFromJson(
     json: JsonString
@@ -34,25 +35,5 @@ export class OutOfBandInvitation {
     const from = jsonObject.from;
 
     return new PrismOnboardingInvitation(onboardingEndpoint, from, type);
-  }
-
-  static parseOutOfBandInvitationFromJson(
-    json: JsonString
-  ): OutOfBandInvitation {
-    const jsonObject = JSON.parse(json);
-    if (!jsonObject.onboardEndpoint) {
-      throw new AgentError.InvitationIsInvalidError(
-        "Undefined PrismOnboardingInvitation onboardEndpoint"
-      );
-    }
-    if (!jsonObject.type) {
-      throw new AgentError.InvitationIsInvalidError(
-        "Undefined PrismOnboardingInvitation type"
-      );
-    }
-    const id = jsonObject.id;
-    const body = jsonObject.body;
-
-    return new OutOfBandInvitation(body, jsonObject.from, id);
   }
 }
